gateway/requirements.txt
flask
grpcio
grpcio-tools
prometheus-client
redis
pika